futurize -f libfuturize.fixes.fix_division_safe

+++ instamatic/calibrate/calibrate_beamshift.py
@@ -2,6 +2,7 @@
 
 from __future__ import print_function
 from __future__ import absolute_import
+from __future__ import division
 import sys, os
 import numpy as np
 import matplotlib.pyplot as plt
@@ -165,7 +166,7 @@ def calibrate_beamshift_live(ctrl, gridsize=None, stepsize=None, save_images=Fal
     shifts = []
     beampos = []
     
-    n = (gridsize - 1) / 2 # number of points = n*(n+1)
+    n = int((gridsize - 1) / 2)  # number of points = n*(n+1)
     x_grid, y_grid = np.meshgrid(np.arange(-n, n+1) * stepsize, np.arange(-n, n+1) * stepsize)
     tot = gridsize*gridsize
