+++ instamatic/calibrate/calibrate_stage_mag1.py
@@ -2,6 +2,7 @@
 
 from __future__ import print_function
 from __future__ import absolute_import
+from __future__ import division
 import sys, os
 import numpy as np
 
@@ -77,7 +78,7 @@ def calibrate_mag1_live(ctrl, gridsize=3, stepsize=2000, save_images=False):
     stagepos = []
     shifts = []
     
-    n = (gridsize - 1) / 2 # number of points = n*(n+1)
+    n = int((gridsize - 1) / 2) # number of points = n*(n+1)
     x_grid, y_grid = np.meshgrid(np.arange(-n, n+1) * stepsize, np.arange(-n, n+1) * stepsize)
     tot = gridsize*gridsize
